EID-828: moved SamlObjectMarshaller and SamlObjectSigner into saml package
To separate SAML processing logic from metadata signing logic
